+++ docs/changelog.md
@@ -3,6 +3,7 @@
 #### 2020 11 28
 * Add option for enable in menu for configuration. This can hide certain options
 * Add Power panel to control power devices via moonraker
+* Add klipper version to system panel
 
 #### 2020 11 18
 * Changed configuration file format.